receive: base logic
