+++ BDKSwiftExampleWallet/BDK Service/BDKSwiftExampleWalletError.swift
