+++ BDKSwiftExampleWallet/Extensions/Int+Extensions.swift
